# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, scheduled, running, paused, finished, cancelled]
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignPage"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{campaignId}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send campaign immediately
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/schedule:
    post:
      operationId: scheduleCampaign
      tags: [campaigns]
      summary: Schedule campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignSchedule"
      responses:
        "200":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/duplicate:
    post:
      operationId: duplicateCampaign
      tags: [campaigns]
      summary: Duplicate a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Get campaign analytics stats
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"

  # ── Templates ────────────────────────────────────────────────────────────
...
  /billing/subscription:
    get:
      operationId: getBillingSubscription
      tags: [billing]
      summary: Get current subscription
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    post:
      operationId: updateBillingSubscription
      tags: [billing]
      summary: Change subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionChange"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, workspaceId, name, type, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        subject:
          type: ["string", "null"]
        previewText:
          type: ["string", "null"]
        fromName:
          type: ["string", "null"]
        fromEmail:
          type: ["string", "null"]
        replyTo:
          type: ["string", "null"]
        type:
          type: string
          enum: [regular, automated, rss, drip, transactional, ab_test, newsletter]
        status:
          type: string
          enum: [draft, scheduled, running, paused, finished, cancelled]
        templateId:
          type: ["integer", "null"]
        listIds:
          type: array
          items:
            type: integer
        segmentIds:
          type: array
          items:
            type: integer
        scheduledAt:
          type: ["string", "null"]
  
        sentAt:
          type: ["string", "null"]
  
        totalRecipients:
          type: integer
        createdAt:
          type: string
  
        updatedAt:
          type: string
    CampaignPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CampaignInput:
      type: object
      required: [workspaceId, name, type]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
          minLength: 1
        subject:
          type: string
        previewText:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string

        replyTo:
          type: string

        type:
          type: string
          enum: [regular, automated, rss, drip, transactional, ab_test, newsletter]
        templateId:
          type: integer
        listIds:
          type: array
          items:
            type: integer
        segmentIds:
          type: array
          items:
            type: integer
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
        previewText:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string

        replyTo:
          type: string

        templateId:
          type: integer
        listIds:
          type: array
          items:
            type: integer
        segmentIds:
          type: array
          items:
            type: integer
    CampaignSchedule:
      type: object
      required: [scheduledAt]
      properties:
        scheduledAt:
          type: string
    CampaignStats:
      type: object
      required: [campaignId, sent, delivered, opened, clicked, bounced, unsubscribed, complained]
      properties:
        campaignId:
          type: integer
        sent:
          type: integer
        delivered:
          type: integer
        opened:
          type: integer
        clicked:
          type: integer
        bounced:
          type: integer
        unsubscribed:
          type: integer
        complained:
          type: integer
        openRate:
          type: number
        clickRate:
          type: number
        bounceRate:
          type: number
        unsubscribeRate:
          type: number
        revenue:
          type: number
        topLinks:
          type: array
          items:
            $ref: "#/components/schemas/LinkStat"
        opensByDevice:
          type: object
          additionalProperties: true
        opensByCountry:
          type: object
          additionalProperties: true
    LinkStat:
      type: object
      required: [url, clicks]
      properties:
        url:
          type: string
        clicks:
          type: integer
        uniqueClicks:
          type: integer

    # ── Templates ──────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, orgId, planId, planName, status, currentPeriodEnd]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
          enum: [active, trialing, past_due, cancelled, incomplete]
        currentPeriodStart:
          type: string
  
        currentPeriodEnd:
          type: string
  
        cancelAtPeriodEnd:
          type: boolean
        subscribersUsed:
          type: integer
        emailsSentThisMonth:
          type: integer
    SubscriptionChange:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
        orgId:
          type: integer